Add galera cluster status check
This PR extends the library to periodically check for galera cluster
health as well.

If node is configured to run in galera cluster (variable `wsrep_on` is
true) but the node is not healthy (status `wsrep_ready` is false) we
omit exclude this node from taking master or slave role.

Main config struct is extended with `SkipGaleraCheck` flag to optionally
disable galera checks.

Minor changes to the tests:

* Test DBs use empty root password, without this spinning galera cluster
on default mariadb docker image fails.
* Extra flag to skip time-zone data loading is added, this greatly
reduces time required to assemble galera cluster.

Approved-by: Algirdas Matas <algirdas.matas@adv.bet>
Approved-by: Dainius Grėbliauskas <dainius.grebliauskas@adv.bet>
Approved-by: Deividas Dunda <deividas.dunda@adv.bet>